+++ js/constants.js
@@ -1,0 +1,27 @@
+const PHOTO_DESCRIPTION_COUNT = 25;
+const PHOTO_DESCRIPTIONS = [
+  'Мгновение',
+  'Настроение',
+  'Деталь',
+  'Эмоции',
+  'История'
+];
+
+const TEXT_MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const COMMENTATOR_NAMES = [
+  'Иван',
+  'Алёна',
+  'Тимофей',
+  'Александр',
+  'Юлия'
+];
+
+export { PHOTO_DESCRIPTION_COUNT, PHOTO_DESCRIPTIONS, TEXT_MESSAGES, COMMENTATOR_NAMES };
