+++ js/utils/generate-id.js
@@ -1,0 +1,6 @@
+import { createRandomNumberFromRangeGenerator } from './random.js';
+
+const generatePhotoId = createRandomNumberFromRangeGenerator(1, 25);
+const generatePhotoName = createRandomNumberFromRangeGenerator(1, 25);
+
+export { generatePhotoId, generatePhotoName };
